+++ src/main/php/ValueReader.php
@@ -11,6 +11,7 @@ namespace stubbles\input;
 use stubbles\date\Date;
 use stubbles\date\span\Day;
 use stubbles\date\span\Month;
+use stubbles\input\errors\ParamErrors;
 use stubbles\input\filter\ArrayFilter;
 use stubbles\input\filter\PasswordFilter;
 use stubbles\input\filter\range\DateRange;
@@ -446,7 +447,7 @@ class ValueReader
         $path = ((null != $basePath) ? ($basePath . '/') : (''));
         return $this->withValidator(new validator\FileValidator($basePath),
                                     'FILE_INVALID',
-                                    ['PATH' => $path . $this->param->getValue()],
+                                    ['PATH' => $path . $this->param->value()],
                                     $default
         );
     }
@@ -470,7 +471,7 @@ class ValueReader
         $path = ((null != $basePath) ? ($basePath . '/') : (''));
         return $this->withValidator(new validator\DirectoryValidator($basePath),
                                     'DIRECTORY_INVALID',
-                                    ['PATH' => $path . $this->param->getValue()],
+                                    ['PATH' => $path . $this->param->value()],
                                     $default
         );
     }
@@ -509,7 +510,7 @@ class ValueReader
     {
         if ($this->param->isNull()) {
             if ($this->required) {
-                $this->paramErrors->add(new ParamError($this->requiredErrorId), $this->param->getName());
+                $this->paramErrors->append($this->param->name(), $this->requiredErrorId);
                 return null;
             }
 
@@ -534,7 +535,7 @@ class ValueReader
     public function withFilter(Filter $filter)
     {
         if ($this->required && $this->param->isEmpty()) {
-            $this->paramErrors->add(new ParamError($this->requiredErrorId), $this->param->getName());
+            $this->paramErrors->append($this->param->name(), $this->requiredErrorId);
             return null;
         }
 
@@ -557,8 +558,8 @@ class ValueReader
             return $value;
         }
 
-        foreach ($this->param->getErrors() as $error) {
-            $this->paramErrors->add($error, $this->param->getName());
+        foreach ($this->param->errors() as $error) {
+            $this->paramErrors->append($this->param->name(), $error);
         }
 
         return null;
@@ -595,8 +596,8 @@ class ValueReader
             return $value;
         }
 
-        foreach ($this->param->getErrors() as $error) {
-            $this->paramErrors->add($error, $this->param->getName());
+        foreach ($this->param->errors() as $error) {
+            $this->paramErrors->append($this->param->name(), $error);
         }
 
         return null;
@@ -612,6 +613,6 @@ class ValueReader
      */
     public function unsecure()
     {
-        return $this->param->getValue();
+        return $this->param->value();
     }
 }