+++ src/main/php/console/BaseConsoleRequest.php
@@ -9,10 +9,10 @@
  */
 namespace stubbles\input\console;
 use stubbles\input\AbstractRequest;
-use stubbles\input\ParamErrors;
 use stubbles\input\Params;
 use stubbles\input\ValueReader;
 use stubbles\input\ValueValidator;
+use stubbles\input\errors\ParamErrors;
 /**
  * Request implementation for command line.
  *
@@ -70,7 +70,7 @@ class BaseConsoleRequest extends AbstractRequest implements ConsoleRequest
      */
     public function getEnvNames()
     {
-        return $this->env->getNames();
+        return $this->env->names();
     }
 
     /**