+++ src/main/php/filter/DayFilter.php
@@ -42,9 +42,9 @@ class DayFilter implements Filter
         }
 
         try {
-            return new Day($param->getValue());
+            return new Day($param->value());
         } catch (IllegalArgumentException $iae) {
-            $param->addErrorWithId('DATE_INVALID');
+            $param->addError('DATE_INVALID');
         }
 
         return null;