+++ src/main/php/web/BaseWebRequest.php
@@ -10,10 +10,10 @@
 namespace stubbles\input\web;
 use stubbles\input\AbstractRequest;
 use stubbles\input\Param;
-use stubbles\input\ParamErrors;
 use stubbles\input\Params;
 use stubbles\input\ValueReader;
 use stubbles\input\ValueValidator;
+use stubbles\input\errors\ParamErrors;
 use stubbles\lang\exception\RuntimeException;
 use stubbles\peer\MalformedUriException;
 use stubbles\peer\http\HttpUri;
@@ -94,7 +94,7 @@ class BaseWebRequest extends AbstractRequest implements WebRequest
      */
     public function getMethod()
     {
-        return strtoupper($this->headers->getValue('REQUEST_METHOD'));
+        return strtoupper($this->headers->value('REQUEST_METHOD'));
     }
 
     /**
@@ -122,7 +122,7 @@ class BaseWebRequest extends AbstractRequest implements WebRequest
         }
 
         $minor = null;
-        if (1 != sscanf($this->headers->get('SERVER_PROTOCOL')->getValue(), 'HTTP/1.%[01]', $minor)) {
+        if (1 != sscanf($this->headers->get('SERVER_PROTOCOL')->value(), 'HTTP/1.%[01]', $minor)) {
             return null;
         }
 
@@ -137,14 +137,14 @@ class BaseWebRequest extends AbstractRequest implements WebRequest
      */
     public function getUri()
     {
-        $host = $this->headers->getValue('HTTP_HOST');
+        $host = $this->headers->value('HTTP_HOST');
         if (strstr($host, ':') === false) {
-            $host .= ':' . $this->headers->getValue('SERVER_PORT');
+            $host .= ':' . $this->headers->value('SERVER_PORT');
         }
 
         $uri  = (($this->headers->has('HTTPS')) ? ('https') : ('http')) . '://'
               . $host
-              . $this->headers->getValue('REQUEST_URI');
+              . $this->headers->value('REQUEST_URI');
         try {
             return HttpUri::fromString($uri);
         } catch (MalformedUriException $murie) {
@@ -160,7 +160,7 @@ class BaseWebRequest extends AbstractRequest implements WebRequest
      */
     public function getHeaderNames()
     {
-        return $this->headers->getNames();
+        return $this->headers->names();
     }
 
     /**
@@ -178,7 +178,7 @@ class BaseWebRequest extends AbstractRequest implements WebRequest
     /**
      * returns error collection for request headers
      *
-     * @return  stubbles\input\ParamErrors
+     * @return  stubbles\input\errors\ParamErrors
      * @since   1.3.0
      */
     public function headerErrors()
@@ -220,7 +220,7 @@ class BaseWebRequest extends AbstractRequest implements WebRequest
      */
     public function getCookieNames()
     {
-        return $this->cookies->getNames();
+        return $this->cookies->names();
     }
 
     /**
@@ -238,7 +238,7 @@ class BaseWebRequest extends AbstractRequest implements WebRequest
     /**
      * returns error collection for request cookies
      *
-     * @return  stubbles\input\ParamErrors
+     * @return  stubbles\input\errors\ParamErrors
      * @since   1.3.0
      */
     public function cookieErrors()
@@ -275,7 +275,7 @@ class BaseWebRequest extends AbstractRequest implements WebRequest
     /**
      * returns error collection for request body
      *
-     * @return  ParamErrors
+     * @return  stubbles\input\errors\ParamErrors
      * @since   1.3.0
      */
     public function bodyErrors()